LPD-7012 - Move page attributes shared by all client extensions to a shared EditClientExtensionsPage.
